make Footer component, replace foreach with Promise.all
