+++ src/utils/getFilteredTodos.ts
@@ -1,0 +1,15 @@
+import { FilterKey } from '../types/FilterKey';
+import { Todo } from '../types/Todo';
+
+export function getFilteredTodos(key: FilterKey, todos: Todo[]) {
+  switch (key) {
+    case FilterKey.All:
+      return todos;
+    case FilterKey.Active:
+      return todos.filter(({ completed }) => !completed);
+    case FilterKey.Completed:
+      return todos.filter(({ completed }) => completed);
+    default:
+      return todos;
+  }
+}
